new version

+++ biglinux-webapps/usr/share/bigbashview/bcc/apps/biglinux-webapps/webapp-remove.sh
@@ -1,47 +1,28 @@
 #!/usr/bin/env bash
 
-#Translation
-export TEXTDOMAINDIR="/usr/share/locale"
-export TEXTDOMAIN=biglinux-webapps
+ICONDESK="$(grep '^Icon=' $filedesk | sed 's|Icon=||')"
+DESKNAME="$(basename $filedesk)"
+NAMEDESK="$(sed 's|-webapp-biglinux-custom.desktop||' <<< $DESKNAME)"
 
-kdialog --title "BigLinux WebApps" --icon "internet-web-browser" \
-        --yesno $"Você tem certeza que deseja remover este WebApp?"
+USER_DESKTOP="$(xdg-user-dir DESKTOP)"
+FOLDER="$HOME/.bigwebapps/$NAMEDESK"
+DATA_DIR="$(grep '^Exec=' $filedesk | sed 's|.*-dir=||;s| --cl.*||')"
+EPI_DIR="$(grep '^Exec=' $filedesk | cut -d' ' -f3 | sed 's|--profile=||')"
 
-if [ "$?" != "0" ]; then
-    exit
-else
-    NAMEDESK="$(basename -s .desktop "$filedesk" | sed 's|-webapp-biglinux-custom||')"
-    ICONDESK="$(grep "Icon=" $filedesk | sed 's|Icon=||')"
-    DESKNAME="$(grep "Name=" $filedesk | sed 's|Name=||')"
+[ -e "$USER_DESKTOP/$DESKNAME" ] && unlink "$USER_DESKTOP/$DESKNAME"
+[ -e "$ICONDESK" ] && rm "$ICONDESK"
+[ -d "$FOLDER" ] && rm -r "$FOLDER"
+[ -d "$EPI_DIR" ] && rm -r "$EPI_DIR" && rmdir $HOME/.config/org.gnome.Epiphany.WebApp*
+[ -d "$DATA_DIR" ] && rm -r "$DATA_DIR"
+[ "$(grep '.local/bin' $filedesk)" ] && {
+    DESKBIN="$(grep '^Exec=' $filedesk  | sed 's|Exec=||')"
+    rm "$DESKBIN"
+}
+rm "$filedesk"
 
-    if [ "$(grep "firefox$" $filedesk)" != "" ];then
+nohup update-desktop-database -q $HOME/.local/share/applications &
+nohup kbuildsycoca5 &> /dev/null &
 
-        if [ -d $HOME/.bigwebapps/"$NAMEDESK" ]; then
-            rm -r $HOME/.bigwebapps/"$NAMEDESK"
-        fi
-        unlink "$(xdg-user-dir DESKTOP)/$DESKNAME.desktop" &> /dev/null
-        rm "$(grep "Exec=" "$filedesk" | sed 's|Exec=||')"
-        xdg-desktop-menu uninstall "$filedesk"
-        rm "$ICONDESK"
-    else
-        if [ -d $HOME/.bigwebapps/"$NAMEDESK" ]; then
-            rm -r $HOME/.bigwebapps/"$NAMEDESK"
-        fi
-        unlink "$(xdg-user-dir DESKTOP)/$DESKNAME.desktop" &> /dev/null
-        xdg-desktop-menu uninstall "$filedesk"
-        rm "$ICONDESK"
-    fi
-
-    nohup update-desktop-database -q $HOME/.local/share/applications &
-    nohup kbuildsycoca5 &> /dev/null &
-    kdialog --title "BigLinux WebApps" --icon "internet-web-browser" \
-    --yesno $"O WebApp foi removido com sucesso!\nVocê deseja remover outro WebApp?"
-
-    if [ "$?" != "0" ]; then
-        echo "<META http-equiv=\"refresh\" content=\"0;URL=index.sh.htm\">"
-        exit
-    else
-        echo "<META http-equiv=\"refresh\" content=\"0;URL=index-remove.sh.htm\">"
-        exit
-    fi
-fi
+resp="$?"
+echo -n $resp
+exit
